D2D Hack
This is the source code of the external eft cheat, which has gone into the public domain because of stupid people. I changed the overlay and am sharing it with you. I use D2DOverlay & kdmapper. I played with this cheat for a few hours and did not get banned. But still I recommend to change the driver. Good luck